backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi[standard]>=0.132.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.3",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "tensorflow>=2.20.0",
    "tensorflow-keras>=0.1",
    "uvicorn>=0.41.0",
]
